+++ src/main/java/com/nttdata/yankitransfers/service/Calculate.java
@@ -3,6 +3,6 @@ package com.nttdata.yankitransfers.service;
 @FunctionalInterface
 public interface Calculate {
 
-    Float calcular(Float monto, Float saldo);
+    Double calcular(Double monto, Double saldo);
 
 }
